requirements.txt
aiosmtplib==4.0.0
altgraph==0.17.4
astroid==4.0.4
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
customtkinter==5.2.2
darkdetect==0.8.0
dill==0.4.1
et_xmlfile==2.0.0
fonttools==4.56.0
fsspec==2025.5.1
git-filter-repo==2.47.0
idna==3.10
isort==8.0.1
mccabe==0.7.0
MouseInfo==0.1.3
msoffcrypto-tool==5.4.2
numpy==1.26.4
olefile==0.47
openpyxl==3.1.3
packaging==24.2
pandas==2.2.2
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.9.4
PyAutoGUI==0.9.54
pycparser==2.22
pydyf==0.10.0
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pylint==4.0.5
PyMsgBox==1.0.9
pyperclip==1.8.2
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
pysqlite3==0.5.4
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.1
pywin32-ctypes==0.2.3
setuptools==75.8.0
six==1.17.0
tomlkit==0.14.0
tzdata==2025.1
xlrd==2.0.1
zopfli==0.2.3.post1
